이기태
멤버컨트롤러 생성
